requirements.txt
asgiref==3.5.2
Django==3.2.13
django-crispy-forms==1.14.0
django-dotenv==1.4.2
pytz==2022.1
sqlparse==0.4.2
